# cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "API Documentation System Management",
        "termsOfService": "https://github.com/ducvubo",
        "contact": {
            "name": "Vũ Đức Bo",
            "url": "https://github.com/ducvubo",
            "email": "vminhduc8@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "system.management.pg.taphoaictu.id.vn",
    "basePath": "/api/v1",
    "paths": {
        "/equipment-maintenance": {
            "get": {
                "description": "GetAllEquipmentMaintenance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipment Maintenance"
                ],
                "summary": "GetAllEquipmentMaintenance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageIndex",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "EqpMtnName",
                        "name": "EqpMtnName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateEquipmentMaintenance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipment Maintenance"
                ],
                "summary": "CreateEquipmentMaintenance",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateEquipmentMaintenanceDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            },
            "patch": {
                "description": "UpdateEquipmentMaintenance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipment Maintenance"
                ],
                "summary": "UpdateEquipmentMaintenance",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateEquipmentMaintenanceDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/equipment-maintenance/recycle": {
            "get": {
                "description": "GetAllEquipmentMaintenanceRecycle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipment Maintenance"
                ],
                "summary": "GetAllEquipmentMaintenanceRecycle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageIndex",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "EqpMtnName",
                        "name": "EqpMtnName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/equipment-maintenance/restore/{id}": {
            "patch": {
                "description": "RestoreEquipmentMaintenance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipment Maintenance"
                ],
                "summary": "RestoreEquipmentMaintenance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/equipment-maintenance/update-status": {
            "patch": {
                "description": "UpdateEquipmentMaintenanceStatus",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipment Maintenance"
                ],
                "summary": "UpdateEquipmentMaintenanceStatus",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateEquipmentMaintenanceStatusDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/equipment-maintenance/{id}": {
            "get": {
                "description": "FindEquipmentMaintenance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipment Maintenance"
                ],
                "summary": "FindEquipmentMaintenance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            },
            "delete": {
                "description": "DeleteEquipmentMaintenance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equipment Maintenance"
                ],
                "summary": "DeleteEquipmentMaintenance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateEquipmentMaintenanceDto": {
            "type": "object",
            "required": [
                "eqp_mtn_date_reported",
                "eqp_mtn_name"
            ],
            "properties": {
                "eqp_mtn_cost": {
                    "type": "integer"
                },
                "eqp_mtn_date_fixed": {
                    "type": "string"
                },
                "eqp_mtn_date_reported": {
                    "type": "string"
                },
                "eqp_mtn_issue_description": {
                    "type": "string"
                },
                "eqp_mtn_location": {
                    "type": "string"
                },
                "eqp_mtn_name": {
                    "type": "string"
                },
                "eqp_mtn_note": {
                    "type": "string"
                },
                "eqp_mtn_performed_by": {
                    "type": "string"
                },
                "eqp_mtn_reported_by": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateEquipmentMaintenanceDto": {
            "type": "object",
            "required": [
                "eqp_mtn_date_reported",
                "eqp_mtn_id",
                "eqp_mtn_name"
            ],
            "properties": {
                "eqp_mtn_cost": {
                    "type": "integer"
                },
                "eqp_mtn_date_fixed": {
                    "type": "string"
                },
                "eqp_mtn_date_reported": {
                    "type": "string"
                },
                "eqp_mtn_id": {
                    "type": "string"
                },
                "eqp_mtn_issue_description": {
                    "type": "string"
                },
                "eqp_mtn_location": {
                    "type": "string"
                },
                "eqp_mtn_name": {
                    "type": "string"
                },
                "eqp_mtn_note": {
                    "type": "string"
                },
                "eqp_mtn_performed_by": {
                    "type": "string"
                },
                "eqp_mtn_reported_by": {
                    "type": "string"
                }
            }
        },
        "model.UpdateEquipmentMaintenanceStatusDto": {
            "type": "object",
            "required": [
                "eqp_mtn_id",
                "eqp_mtn_status"
            ],
            "properties": {
                "eqp_mtn_id": {
                    "type": "string"
                },
                "eqp_mtn_status": {
                    "description": "ENUM: pending, in_progress, done, rejected",
                    "type": "string"
                }
            }
        },
...
        "response.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "status code",
                    "type": "integer"
                },
                "data": {
                    "description": "dữ liệu trả về"
                },
                "error": {
                    "description": "dữ liệu lỗi"
                },
                "message": {
                    "description": "thông báo lỗi"
                },
                "statusCode": {
                    "description": "HTTP status code",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AccessTokenAuth": {
            "type": "apiKey",
            "name": "x-at-tk",
            "in": "header"
        },
        "ClientIDAuth": {
            "type": "apiKey",
            "name": "id_user_guest",
            "in": "header"
        },
        "RefreshTokenAuth": {
            "type": "apiKey",
            "name": "x-rf-tk",
            "in": "header"
        }
    },
    "security": [
        {
            "ClientIDAuth": []
        },
        {
            "AccessTokenAuth": []
        },
        {
            "RefreshTokenAuth": []
        }
    ]
}